+++ tests/test_package.py
@@ -1,0 +1,7 @@
+from pandas.core.api import DataFrame as DataFrame
+
+import multi_condition_comparisions
+
+
+def test_package_has_version():
+    assert multi_condition_comparisions.__version__ is not None
